error handling for each destroy

+++ Signup-API/app/graphql/mutations/delete_item.rb
@@ -12,10 +12,16 @@ module Mutations
                 if item.signup.present?
                     
                     #if the item has been signed up for... we need to delete it from its User's signup to avoid foreign key violations
-                    u = User.find(item.signup.user_id)
-                    u.signups.find(item.signup.id).destroy
+                    user = User.find(item.signup.user_id)
+                    selected_signup = user.signups.find(item.signup.id)
+                    selected_signup.destroy
+                    raise GraphQL::ExecutionError, selected_signuperrors.full_messages.join(", ") unless selected_signup.errors.empty?
+                   
                     item.signup.destroy
+                    raise GraphQL::ExecutionError, item.signup.errors.full_messages.join(", ") unless item.signup.errors.empty?
+                    
                     item.destroy
+                    raise GraphQL::ExecutionError, item.errors.full_messages.join(", ") unless item.errors.empty?
                     item
 
                 else
